tests data

+++ tasks/kolotukhin_a_elem_vec_sum/common/include/common.hpp
@@ -7,9 +7,10 @@
 
 namespace kolotukhin_a_elem_vec_sum {
 
-using InType = std::vector<int>;
-using OutType = int;
-using TestType = int;
+
+using InType = std::vector<uint64_t>;
+using OutType = uint64_t;
+using TestType = uint64_t;
 using BaseTask = ppc::task::Task<InType, OutType>;
 
 }  // namespace kolotukhin_a_elem_vec_sum
